# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Client API",
        "description": "API documentation for Imperial Client Mobile & Web Applications",
        "contact": {
            "email": "support@imperial.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "List all reviews",
                "operationId": "7d8a0635268917440915cc68c478b000",
                "parameters": [
                    {
                        "name": "club_id",
                        "in": "query",
                        "description": "Filter by Club ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a review for a club",
                "operationId": "63dea222a42904c6783e3fb43e7e8b55",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "club_id"
                                ],
                                "properties": {
                                    "club_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "booking_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "is_anonymous": {
                                        "type": "integer",
                                        "example": 0,
                                        "nullable": true
                                    },
                                    "remark": {
                                        "type": "string",
                                        "example": "you can put some text which from added",
                                        "nullable": true
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Amazing place, great music!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/qrcode/scan": {
            "get": {
                "tags": [
                    "QR Code"
                ],
                "summary": "View booking details by QR code",
                "operationId": "9a3a471cccfc9ca3bd44763cc3810350",
                "parameters": [
                    {
                        "name": "qr_code_id",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details associated with QR code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found for QR code"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/qrcode/booking-checkin": {
            "post": {
                "tags": [
                    "QR Code"
                ],
                "summary": "Validate QR code and check in booking",
                "operationId": "aef25bbd1db2cba70553010ae559d5b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qr_code_id"
                                ],
                                "properties": {
                                    "qr_code_id": {
                                        "type": "string",
                                        "example": "XXXXXXXX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking checked in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking checked in successfully."
                                        },
                                        "booking": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid QR code or Booking already checked in or cancelled"
                    },
                    "403": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum Bearer token",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}